added jstXml, for xml parsing. started implementing GameFileConverter
namespace jst is a custom lib, designed to house lightweight tools. ist XmlParser and xmlNode are designed to be fast, simple and hide complexity in abstraction. GamFileConverter was extended and will be able to Open Avalanche Filetypes directly by invoking Commandline tools like Ashen's or PredatorCZ's to convert to xml and back. this allows "directly" editing avalanche files.